Connect js of forms with templates that use forms
