+++ estimage/plugins/redhat_compliance/__init__.py
@@ -11,13 +11,10 @@ from ...visualize.burndown import StatusStyle
 from .. import jira
 from .forms import AuthoritativeForm, ProblemForm
 
-JiraFooter = jira.JiraFooter
-
 
 EXPORTS = dict(
     AuthoritativeForm="AuthoritativeForm",
     ProblemForm="ProblemForm",
-    Footer="JiraFooter",
     BaseCard="BaseCard",
     MPLPointPlot="MPLPointPlot",
     Statuses="Statuses",
