+++ estimage/plugins/redhat_compliance/templates/rhcompliance.html
@@ -13,4 +13,7 @@
 
 {% endblock %}
 
-
+{% block footer %}
+{{ super() }}
+{{ plugin_form.supporting_js([plugin_form]) | safe }}
+{% endblock %}
