+++ estimage/webapp/templates/issue_view.html
@@ -103,3 +103,12 @@
     {%- endif %}
 </div>
 {% endblock %}
+
+{% block footer %}
+{{ super() }}
+{% if forms %}
+{% if "authoritative" in forms -%}
+{{ forms["authoritative"].supporting_js([forms["authoritative"]]) | safe }}
+{%- endif %}
+{%- endif %}
+{% endblock %}
